Telegram Bot/requirements.txt
Flask==2.1.0
python-telegram-bot==20.0
